// packages/old/src/types/types.ts
export interface IRestaurantsEL {
  facilities: string[];
  restaurants: {
    acceptsOffers: boolean;
    addressLine1: string;
    city: string;
    email: string;
    facilities: string[];
    lab: boolean;
    latitude: number;
    longitude: number;
    mopEnabled: boolean;
    name: string;
    openingHours: {
      categoryName: string;
      hours: {
        end: string;
        start: string;
        status: string;
        weekday: string;
      }[];
    }[];
    phone: string;
    rid: string;
    zipCode: string;
  }[];
}

export enum APIType {
  AP = 'AP',
  EL = 'EL',
  EU = 'EU',
  HK = 'HK',
  UNKNOWN = 'UNKNOWN',
  US = 'US',
}